Add paintr-widgets
