+++ crates/paintr_widgets/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "paintr"
+name = "paintr_widgets"
 version = "0.1.0"
 authors = ["Edwin Cheng <edwin0cheng@gmail.com>"]
 edition = "2018"
@@ -7,8 +7,7 @@ edition = "2018"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-paintr_core = {path="../paintr_core"}
-paintr_widgets = {path="../paintr_widgets"}
+paintr_core = {path = "../paintr_core"}
 
 druid = { version = "0.7.0", features = ["svg"] }
 image = "0.23.12"
