+++ crates/paintr_widgets/src/widgets/notif_bar.rs
