+++ crates/paintr_widgets/src/widgets/painter.rs
